[Ref #351] Reduce SearchDao to FacetedSearchDao removing full text search without lucene.
Removes graphdb maven profile and lucene Spring profile
