+++ src/main/java/cz/cvut/kbss/termit/service/init/SystemInitializer.java
@@ -17,7 +17,7 @@
  */
 package cz.cvut.kbss.termit.service.init;
 
-import cz.cvut.kbss.termit.service.init.lucene.LuceneConnectorInitializer;
+import cz.cvut.kbss.termit.service.init.lucene.GraphDBLuceneConnectorInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.SmartInitializingSingleton;
@@ -47,7 +47,7 @@ public class SystemInitializer implements SmartInitializingSingleton, Ordered {
         appContext.getBean(VocabularyPrimaryLanguageGenerator.class).generateMissingPrimaryLanguage();
         appContext.getBean(VocabularyAccessControlListGenerator.class).generateMissingAccessControlLists();
         appContext.getBean(VocabularyAnonymousAccessControlListGenerator.class).generateMissingAccessControlLists();
-        appContext.getBean(LuceneConnectorInitializer.class).initialize();
+        appContext.getBean(GraphDBLuceneConnectorInitializer.class).initialize();
     }
 
     @Override
