--- src/main/java/cz/cvut/kbss/termit/service/init/lucene/LuceneConnectorInitializer.java
@@ -1,8 +1,0 @@
-package cz.cvut.kbss.termit.service.init.lucene;
-
-/**
- * Object responsible for initializing Lucene connectors in the GraphDB
- */
-public interface LuceneConnectorInitializer {
-    void initialize();
-}
